New loader interfaces
