+++ modules/core/src/index.js
@@ -54,6 +54,8 @@ export {mergeXVIZStyles} from './utils/style';
 // Constants
 export {COORDINATE, VIEW_MODE} from './constants';
 
+export {default as _LoaderInterface} from './loaders/loader-interface';
+export {default as _PlayableLoaderInterface} from './loaders/playable-loader-interface';
 export {default as _XVIZLoaderInterface} from './loaders/xviz-loader-interface';
 export {default as XVIZStreamLoader} from './loaders/xviz-stream-loader';
 export {default as XVIZLiveLoader} from './loaders/xviz-live-loader';
