+++ modules/core/src/loaders/xviz-loader-interface.js
@@ -21,21 +21,17 @@
 import {getXVIZConfig, StreamSynchronizer, LOG_STREAM_MESSAGE} from '@xviz/parser';
 import {clamp} from 'math.gl';
 
+import PlayableLoaderInterface from './playable-loader-interface';
 import createSelector from '../utils/create-selector';
 import stats from '../utils/stats';
 
 /* eslint-disable callback-return */
-export default class XVIZLoaderInterface {
+export default class XVIZLoaderInterface extends PlayableLoaderInterface {
   constructor(options = {}) {
+    super();
     this.options = options;
     this._debug = options.debug || (() => {});
     this.callbacks = {};
-
-    this.listeners = [];
-    this.state = {};
-    this._updates = 0;
-    this._version = 0;
-    this._updateTimer = null;
   }
 
   /* Event types:
@@ -69,32 +65,6 @@ export default class XVIZLoaderInterface {
       }
     }
     stats.get(`loader-${eventType}`).incrementCount();
-  }
-
-  subscribe(instance) {
-    this.listeners.push(instance);
-  }
-
-  unsubscribe(instance) {
-    const index = this.listeners.findIndex(o => o === instance);
-    if (index >= 0) {
-      this.listeners.splice(index, 1);
-    }
-  }
-
-  get(key) {
-    return this.state[key];
-  }
-
-  set(key, value) {
-    if (this.state[key] !== value) {
-      this.state[key] = value;
-      this._version++;
-      if (!this._updateTimer) {
-        /* global requestAnimationFrame */
-        this._updateTimer = requestAnimationFrame(this._update);
-      }
-    }
   }
 
   onXVIZMessage = message => {
@@ -232,17 +202,6 @@ export default class XVIZLoaderInterface {
       return null;
     }
   );
-
-  /* Private actions */
-  _update = () => {
-    this._updateTimer = null;
-    this.listeners.forEach(o => o(this._version));
-  };
-
-  _bumpDataVersion() {
-    this._updates++;
-    this.set('dataVersion', this._updates);
-  }
 
   /* Subclass hooks */
